requirements.txt
google-cloud-storage==2.18.2
google-cloud-aiplatform==1.68.0